chore: extract preconfigured_policy steps to its file
